+++ config/install.yaml
@@ -2517,6 +2517,7 @@ data:
         - kind: InterstepBufferService
           schedule: "0,0,10"
       analysisRunTimeout: 1200
+    permittedRiders: "group=autoscaling.k8s.io,kind=VerticalPodAutoscaler"
 kind: ConfigMap
 metadata:
   name: numaplane-controller-config
