+++ config/samples/numaplane.numaproj.io_v1alpha1_monovertexrollout.yaml
@@ -12,14 +12,6 @@ spec:
       templates:
   riders:
     - definition:
-        apiVersion: v1
-        kind: ConfigMap
-        metadata:
-          name: source-configmap
-        data:
-          something: something
-      progressive: true
-    - definition:
         apiVersion: autoscaling.k8s.io/v1beta2
         kind: VerticalPodAutoscaler
         metadata: 
@@ -29,6 +21,7 @@ spec:
             apiVersion: numaproj.io/v1alpha1
             kind: MonoVertex
             name: '{{.monovertex-name}}'
+          # todo: add more fields here...
 
   monoVertex:
     #uncomment for Progressive rollout to set Numaflow Controller instance:
@@ -44,9 +37,6 @@ spec:
         udsource:
           container:
             image: quay.io/numaio/numaflow-go/source-simple-source:stable
-            volumeMounts:
-              - name: source-cm-volume
-                mountPath: /etc/config
         transformer:
           container:
             image: docker.intuit.com/quay-rmt/numaio/numaflow-rs/source-transformer-now:stable
@@ -54,7 +44,3 @@ spec:
         udsink:
           container:
             image: quay.io/numaio/numaflow-go/sink-log:stable
-      volumes:
-        - name: source-cm-volume
-          configMap:
-            name: source-configmap-{{.monovertex-name}}
