+++ internal/controller/isbservicerollout/isbservicerollout_controller.go
@@ -1055,7 +1055,7 @@ func (r *ISBServiceRolloutReconciler) GetDesiredRiders(rolloutObject ctlrcommon.
 	}
 
 	// verify that desiredRiders are all permitted Kinds
-	if !riders.VerifyRiders(desiredRiders) {
+	if !riders.VerifyRidersPermitted(desiredRiders) {
 		return desiredRiders, fmt.Errorf("rider definitions contained unpermitted Kind")
 	}
