+++ internal/controller/pipelinerollout/pipelinerollout_controller.go
@@ -1365,7 +1365,7 @@ func (r *PipelineRolloutReconciler) GetDesiredRiders(rolloutObject ctlrcommon.Ro
 	}
 
 	// verify that desiredRiders are all permitted Kinds
-	if !riders.VerifyRiders(desiredRiders) {
+	if !riders.VerifyRidersPermitted(desiredRiders) {
 		return desiredRiders, fmt.Errorf("rider definitions contained unpermitted Kind")
 	}
